Give a URL as an example for this input
This 'Sign in' form, at least in Chrome, is picky about input - trying macwright.org is caught by form validation, but adding https:// works. There's the other route - using a text form instead of a URL form. But if the goal is to use a URL form, the example should be a URL so that folks don't get tripped up by trying to use a domain.